fix linting issues and org
