+++ src/app/financial-chart/panes/financial-chart-panes.component.html
@@ -1,4 +1,4 @@
-<igx-financial-chart height="500px" width="100%"
+<igx-financial-chart height="100%" width="100%"
 	[dataSource]="data" 
 	#chart>
 </igx-financial-chart>
