+++ src/app/financial-chart/performance/financial-chart-performance.component.ts
@@ -1,20 +1,18 @@
-/* tslint:disable:no-trailing-whitespace max-line-length */
-
 import { ChangeDetectionStrategy, Component } from "@angular/core";
 import { StockDataService } from "../services/stock-data.service";
 
 @Component({
     changeDetection: ChangeDetectionStrategy.OnPush,
     providers: [ StockDataService ],
-    selector:  "app-financial-chart-titles",
-    styleUrls: [ "./financial-chart-titles.component.scss"],
-    templateUrl: "./financial-chart-titles.component.html"
+    selector:  "app-financial-chart-performance",
+    styleUrls: [ "./financial-chart-performance.component.scss"],
+    templateUrl: "./financial-chart-performance.component.html"
 })
-export class FinancialChartTitlesComponent {
+export class FinancialChartPerformanceComponent {
 
     public data: any;
 
     constructor(private dataService: StockDataService) {
         this.data = [ this.dataService.GetStockTSLA() ];
     }
-} 
+}
