+++ src/app/financial-chart/titles/financial-chart-titles.component.ts
@@ -1,16 +1,14 @@
-/* tslint:disable:no-trailing-whitespace max-line-length */
-
 import { ChangeDetectionStrategy, Component } from "@angular/core";
 import { StockDataService } from "../services/stock-data.service";
 
 @Component({
     changeDetection: ChangeDetectionStrategy.OnPush,
     providers: [ StockDataService ],
-    selector:  "app-financial-chart-performance",
-    styleUrls: [ "./financial-chart-performance.component.scss"],
-    templateUrl: "./financial-chart-performance.component.html"
+    selector:  "app-financial-chart-titles",
+    styleUrls: [ "./financial-chart-titles.component.scss"],
+    templateUrl: "./financial-chart-titles.component.html"
 })
-export class FinancialChartPerformanceComponent {
+export class FinancialChartTitlesComponent {
 
     public data: any;
